requirements.txt
# Core streaming dependencies
pyrogram==2.0.106
telethon==1.38.1
tgcrypto==1.2.5
# pytgcalls is not installable on Windows via PyPI in some releases (depends on `tgcalls` package).
# Install pytgcalls later on Linux / WSL or from source. See README.
# pytgcalls==2.1.0
yt-dlp==2025.1.12
ffmpeg-python==0.2.0
pydantic==2.9.2
python-dotenv==1.0.1

# Prometheus metrics for production monitoring
prometheus_client==0.16.0

# Analysis tool dependencies
radon==6.0.1
pylint==3.0.3
bandit==1.7.5
pydocstyle==6.3.0
# safety and pip-audit have conflicting packaging requirements on some platforms.
# Install them separately in a dev environment if needed.
# safety==2.3.5
# pip-audit==2.6.1
jinja2==3.1.2
markdown==3.5.1
